chore: 패키지 구조 변경 및 테이블 이름 지정
