+++ src/main/java/GGUM_Team3/Server/domain/auth/dto/request/MailSendRequest.java
@@ -2,10 +2,12 @@ package GGUM_Team3.Server.domain.auth.dto.request;
 
 import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotEmpty;
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 @AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
+@NoArgsConstructor
 @Getter
 public class MailSendRequest {
     @Email
