+++ src/main/java/GGUM_Team3/Server/domain/like/entity/LikeEntity.java
@@ -1,0 +1,31 @@
+package GGUM_Team3.Server.domain.like.entity;
+
+import GGUM_Team3.Server.domain.user.entity.UserEntity;
+import GGUM_Team3.Server.global.BaseEntity;
+import GGUM_Team3.Server.domain.meeting.entity.Meeting;
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@Table(name = "like")
+public class LikeEntity extends BaseEntity {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private Long id;
+
+    @ManyToOne
+    @JoinColumn(name = "meeting_id", nullable = false)
+    private Meeting meeting;
+
+    @ManyToOne
+    @JoinColumn(name = "user_id", nullable = false)
+    private UserEntity user;
+}
