+++ src/main/java/GGUM_Team3/Server/domain/tag/category/dto/CategoryDTO.java
@@ -1,0 +1,16 @@
+package GGUM_Team3.Server.domain.tag.category.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class CategoryDTO {
+    private int categoryId;
+    private String categoryName;
+
+}
