+++ src/main/java/GGUM_Team3/Server/domain/tag/hashtag/dto/HashtagDTO.java
@@ -1,0 +1,17 @@
+package GGUM_Team3.Server.domain.tag.hashtag.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.UUID;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class HashtagDTO {
+    private UUID hashtagId;
+    private String hashtagName;
+}
