+++ src/main/java/GGUM_Team3/Server/domain/tag/hashtag/repository/MeetingHashtagRepository.java
@@ -1,0 +1,16 @@
+package GGUM_Team3.Server.domain.tag.hashtag.repository;
+
+import GGUM_Team3.Server.domain.tag.hashtag.entity.MeetingHashtagEntity;
+import GGUM_Team3.Server.domain.meeting.entity.Meeting;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.List;
+import java.util.UUID;
+
+@Repository
+public interface MeetingHashtagRepository extends JpaRepository<MeetingHashtagEntity, UUID> {
+    List<MeetingHashtagEntity> findByHashtagEntity_HashtagName(String hashtagName);
+
+    List<MeetingHashtagEntity> findByMeeting(Meeting meeting);
+}
