--- src/main/java/GGUM_Team3/Server/like/entity/LikeEntity.java
@@ -1,29 +1,0 @@
-package GGUM_Team3.Server.like.entity;
-
-import GGUM_Team3.Server.domain.user.entity.UserEntity;
-import GGUM_Team3.Server.meeting.entity.Meeting;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class LikeEntity {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Long id;
-
-    @ManyToOne
-    @JoinColumn(name = "meeting_id", nullable = false)
-    private Meeting meeting;
-
-    @ManyToOne
-    @JoinColumn(name = "user_id", nullable = false)
-    private UserEntity user;
-}
